+++ CHANGELOG.md
@@ -35,3 +35,5 @@ This project tries to follow [SemVer 2.0.0](https://semver.org/).
   as fast, low memory using, extensible, and highly customizable logging
   libraries. Heavily inspired by [rs/zerolog](https://github.com/rs/zerolog).
   (#10, #13, #15)
+
+- Added new error response functions. (#17)